huggingface/pytorch/tgillamacpp/docker/tgi-llamacpp-requirements.txt
boto3
dataclasses
docker
gitpython
sagemaker

parameterized
pytest
pytest-mock
pytest-xdist